// client/src/model/apartment.ts
export class Apartment {
  id: number = 0;
  hasWater: boolean = false;
  hasGas: boolean = false;
  hasElectricity: boolean = false;
  hasSewer: boolean = false;
  condition: string = "";
  walls: string = "";
  hasHotWater: boolean = false;
  hasHeating: boolean = false;
  bath: string = "";
  toilet: string = "";
  constructionTime: string = "";
  hasLoggia: boolean = false;
  hasBalcony: boolean = false;
  area: number = 0;
  roomCount: number = 0;
  levelCount: number = 0;
  level: number = 0;
  ceiling: number = 0;
  hasGarage: boolean = false;
  hasGarbageChute: boolean = false;
  hasElevator: boolean = false;
  isCorner: boolean = false;
}